+++ CHANGELOG.md
@@ -7,6 +7,9 @@ NOTE: all releases may include dependency updates, not specifically mentioned
 ## UNRELEASED
 
 - feat: integrate try-as library [#912](https://github.com/hypermodeinc/modus/pull/912)
+
+## 2025-07-12 - Runtime v0.18.5
+
 - fix: sentry source context [#940](https://github.com/hypermodeinc/modus/pull/940)
 - fix: connect kubernetes logger [#945](https://github.com/hypermodeinc/modus/pull/945)
 - fix: don't wait indefinitely for k8s secrets [#947](https://github.com/hypermodeinc/modus/pull/947)